backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-socketio[asyncio]==5.12.1
anthropic==0.52.0
stripe==12.2.0
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.1
python-dotenv==1.1.0
